Treat responses as binary
The encoding looks to be latin1 and not utf8, so decoding them as utf8 causes
errors. BeautifulSoup handles binary data just fine so we can save ourselves the
work of having to figure out the encoding and decode it.

This means we can't do a simple string-replace for the typo fixes, we'll have to
either marshal the data or do the replacements after beautifulsoup extracts the
text.