WIP scraper fixes (#119)
* Use https for current pages

* Drop RATB from agencies

Currently www.foia.gov returns an error for this agency. We'll have to follow up
with them.

* Treat responses as binary

The encoding looks to be latin1 and not utf8, so decoding them as utf8 causes
errors. BeautifulSoup handles binary data just fine so we can save ourselves the
work of having to figure out the encoding and decode it.

This means we can't do a simple string-replace for the typo fixes, we'll have to
either marshal the data or do the replacements after beautifulsoup extracts the
text.

* Remove unsupported pip option

* Parsing fix

FLRA The Authority, Office of Case Intake and Publication has very empty data.
Just skip parsing the address.

* Add some assertions for downloading agency data

- Check HTTP status
- Check for Request Rejected message
- Assume latin1 encoding, then encode back to utf8
- Warn when agencies fail to download rather than crash
- Re-enable typos
- Re-enable RATB (it still fails to download)