Fix the scrapers and update the latest data (#120)
* Replace http -> https

www.foia.gov moved to https and redirects any http requests to the home page.
Without a 404, it's difficult to detect this. Globally replace www.foia.gov urls
fixes most issues.

* Add assertions for HTTP responses

* Update the data files

* Fix BeautifulSoup warnings

Be explicit about the parser to use.

* Alphabetize list of agencies

* Update list of agencies

* Add Makefiles for common contacts tasks

This is incomplete, doesn't touch the other subdirectories like documents.

* Add data for "new" agencies

* Remove obsolete manual data

This manual data was curated by us in 2015. It's a tax we carry. We have to
maintain this information by keeping it up to date. We (conservatively) removed
what could be successfully scraped.

Methodology:
1. Remove all the `manual_data/`
2. Run the full scraper process
3. Diff the `data/`
4. Look at each agency that changed
   a. If information was missing, we restored the `manual_data` for the agency.
   b. If information was changed, we checked to make sure it was valid. For
      links, we checked they went to the correct resource. Names, addresses, and
      emails were only validated that they _appeared_ to be valid data.
   c. The restored `manual_data` was updated to reflect the good data.
5. Run the full scraper process to verify

In some cases, both the old manual data was wrong, and the scraped data was
wrong. We went with the scraped data in those cases and made a little effort to
see if the correct data could be found online. In a few cases, we couldn't find
correct data, so the incorrect scraped data remains.

* Update data

* Note about USAGov contacts